backend/requirements.txt
fastapi
python-multipart
uvicorn[standard]
sqlalchemy
psycopg2-binary
numpy
faiss-cpu
email-validator
alembic
pydantic-settings
python-jose[cryptography]
bcrypt
sentence-transformers
langgraph
langchain-core
langchain-groq
resend
